+++ protocol/asetekpro/core.h
@@ -29,7 +29,7 @@ enum AsetekProSpeed {
 };
 
 int corsairlink_asetekpro_firmware_id(struct corsair_device_info *dev,
-            struct libusb_device_handle *handle, char *firmware, size_t firmware_size);
+            struct libusb_device_handle *handle, char *firmware, uint8_t firmware_size);
 
 /* fan */
 int corsairlink_asetekpro_fan_speed(struct corsair_device_info *dev,
